% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculate_area}
\alias{calculate_area}
\title{Area between curves (ABC)}
\usage{
calculate_area(x, lower, upper, ...)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{lower}{the lower limit of the integral}

\item{xlim}{slice the data. Default \code{NULL} means the data isn't slice.}

\item{...}{passed to wellz::group}
}
\description{
Calculate the area under each curve and return a data
frame with the metadata for each well (i.e., the compounds,
cell type, and concentration of compounds in the well). This
is essentially a wrapper around \code{wellz::area_under_curve}
}

