% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculate_max_rate}
\alias{calculate_max_rate}
\title{Maximum slope of curve (MaxS)}
\usage{
calculate_max_rate(x, ...)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{...}{passed to wellz::max_rate}
}
\description{
Calculate the maximum slope of a curve using \code{wellz::max_rate}.
The results are returned in a data frame with the rate for each
well and the well metadata in columns (i.e., the compound, cell type,
and concentration of compounds in each well).
}

