% Generated by roxygen2 (4.0.2): do not edit by hand
\name{figure_2}
\alias{figure_2}
\title{Make Figure 2}
\usage{
figure_2(wells = NULL)
}
\arguments{
\item{wells}{the output of \code{load_data}}
}
\description{
Makes the different panels for Figure 2 and combines them.
The wells object is loaded with \code{load_data} if it is
not provided. Before the figure is printed, two multi-panel
plots are also printed showing the area between curves (ABC)
and maximum slow (MaxS) for different concentrations and cell types.
}

