% Generated by roxygen2 (4.0.2): do not edit by hand
\name{figure_3}
\alias{figure_3}
\title{Make Figure 3}
\usage{
figure_3(wells = NULL)
}
\arguments{
\item{wells}{the output of \code{load_data}}
}
\description{
Makes the different panels for Figure 3 and combines them.
The wells object is loaded with \code{load_data} if it is
not provided. Panel B was made manually by piecing together
different image files in Adobe Illustrator. The data for the
bar plots is in the \code{"./inst/extdata/CCK8"} directory and
the annotations of the wells is in \code{"./inst/extdata/CCK8/Annotations.csv"}.
}

