% Generated by roxygen2 (4.0.2): do not edit by hand
\name{figure_4}
\alias{figure_4}
\title{Make Figure 4}
\usage{
figure_4(wells = NULL)
}
\arguments{
\item{wells}{the output of \code{load_data}}
}
\description{
Makes the different panels for Figure 4 and combines them.
The wells object is loaded with \code{load_data} if it is
not provided.
}

