% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_data}
\alias{load_data}
\title{load raw data from text files}
\usage{
load_data()
}
\description{
This uses the metadata in \code{"extdata/Annotations.csv"}
to make a \code{wellList} object. The data files are in the
\code{"extdata"} directory of the R package. Use \code{find.package}
to find the package directory.
}

