% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalize_toxin}
\alias{normalize_toxin}
\title{Custom normalization}
\usage{
normalize_toxin(x, ID = "toxinAdd", xlim = NULL, ...)
}
\arguments{
\item{x}{a \code{wellList} object}

\item{ID}{a \code{character}}

\item{xlim}{slice the data. Default \code{NULL} means the data isn't slice.}

\item{...}{passed to wellz::transform}
}
\description{
Normalizes impedance curves. By default, \code{ID="toxinAdd"}, so
that the data will be centered and normalized to the time at which
toxin was added
}

