% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.wellList}
\alias{plot.wellList}
\title{change defaults of wellz::plot.wellList}
\usage{
\method{plot}{wellList}(..., color = "concentration", replicates = TRUE,
  ID = "toxinAdd", title = NULL)
}
\description{
The new defaults are given in the example call. If
\code{title} is not \code{NULL}, then it must be a string
that can be used with \code{ggtitle}
}

