% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select}
\alias{select}
\title{Change default of wellz::select}
\usage{
select(..., controls = TRUE)
}
\arguments{
\item{...}{passed to \code{wellz::select}}
}
\description{
Changes the default arguments of some of the arguments of
\code{wellz::select}: \code{controls=TRUE}.
}

