% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select.wellList}
\alias{select.wellList}
\title{Change default of wellz::select}
\usage{
\method{select}{wellList}(..., controls = TRUE)
}
\arguments{
\item{...}{passed to \code{wellz::select}}
}
\description{
Changes the default arguments of some of the arguments of
\code{wellz::select}: \code{controls=TRUE}.
}

